Warning Pop Up Msg
Created warning pop up message. When player is near an obstacle, message will pop up. Player has the option to turn it off (that is, the message won't display every single time there is an obstacle). However, I need to figure out how to displays the warning only once for said obstacle. Right now, it looks like a glitch if the player tries to close the pop up.