Quick mapCam fix
So I haven't been able to figure out how to take off the map from the screen when the map toggle is off. However, a quick fix is just to deactivate the camera and display an empty screen in its place. It works in game, but when built the image is moved. Need to fix.