Fixed Height Moving Bug
The player would not be able to move forward/backward if there was an object in the way. However, when I implemented the height slider the user at a different height from the original height could go over objects. I have fixed that issue. From any height, if there is an obstacle in the front or back the user cannot pass through it.